backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
google-api-python-client>=2.122.0
google-auth-oauthlib>=1.2.0
groq>=0.4.2
sqlalchemy>=2.0.28
numpy>=1.26.4
python-dotenv>=1.0.1
pydantic>=2.6.4
pydantic-settings>=2.2.1
